split StackIterator in its own file
