+++ src/som/primitives/SystemPrims.java
@@ -61,8 +61,8 @@ import som.vmobjects.SObjectWithClass;
 import som.vmobjects.SSymbol;
 import tools.SourceCoordinate;
 import tools.asyncstacktraces.ShadowStackEntry;
-import tools.asyncstacktraces.ShadowStackEntry.HaltStackIterator;
 import tools.asyncstacktraces.ShadowStackEntryLoad;
+import tools.asyncstacktraces.StackIterator.HaltStackIterator;
 import tools.concurrency.TraceParser;
 import tools.concurrency.TracingActors.TracingActor;
 import tools.concurrency.TracingBackend;
